mysql instead of sqlite3
openproject doesn't support sqlite, see also
https://github.com/opf/openproject/pull/2204

+++ .sandstorm/unprivileged-setup.sh
@@ -2,33 +2,44 @@
 set -euo pipefail
 
 # {{{ ruby setup
-cd /opt/ruby
-
-git clone --depth 1 https://github.com/rbenv/rbenv.git /opt/ruby/rbenv
-git clone --depth 1 https://github.com/rbenv/ruby-build.git /opt/ruby/rbenv/plugins/ruby-build
+if [ ! -e /opt/ruby/rbenv ]
+then
+  git clone --depth 1 https://github.com/rbenv/rbenv.git /opt/ruby/rbenv
+  git clone --depth 1 https://github.com/rbenv/ruby-build.git /opt/ruby/rbenv/plugins/ruby-build
+fi
 
 export PATH=/opt/ruby/rbenv/bin:$PATH
 export RBENV_ROOT=/opt/ruby/rbenv
 eval "$(rbenv init -)"
 
-rbenv install 2.3.0
-rbenv shell 2.3.0
+rbenv shell 2.3.0 || {
+  rbenv install 2.3.0
+  rbenv shell 2.3.0
+}
 
-gem install bundler
+if [ ! -f "$(which bundle)" ]
+then
+  gem install bundler
+fi
 
 # }}}
 
 # {{{ nodejs setup
 
-git clone --depth 1 https://github.com/OiNutter/nodenv.git /opt/node/nodenv
-git clone --depth 1 https://github.com/OiNutter/node-build.git /opt/node/nodenv/plugins/node-build
+if [ ! -e /opt/node/nodenv ]
+then
+  git clone --depth 1 https://github.com/OiNutter/nodenv.git /opt/node/nodenv
+  git clone --depth 1 https://github.com/OiNutter/node-build.git /opt/node/nodenv/plugins/node-build
+fi
 
 export PATH=/opt/node/nodenv/bin:$PATH
 export NODENV_ROOT=/opt/node/nodenv
 eval "$(nodenv init -)"
 
-nodenv install 0.12.7
-nodenv shell 0.12.7
+nodenv shell 0.12.7 || {
+  nodenv install 0.12.7
+  nodenv shell 0.12.7
+}
 
 # }}}
